Refactor addUserXP and add removeUserXP function
Refactor the addUserXP function in database.js to include a new removeUserXP function. This function allows for the removal of XP from a user in a guild. The removeUserXP function checks if the user is in the guild's levels and if their XP is greater than the amount to remove. If so, it decreases their XP by the specified amount and updates the guild's levels in the database. The updatedGuild variable is returned to confirm the successful removal of XP.

Add addxp command to add XP to a user

Add a new command, addxp.js, in the commands/levels directory. This command allows for the addition of XP to a specified user. It takes two options: the user to add XP to and the amount of XP to add. The execute function retrieves the guild's bot colours and the user and XP options from the interaction. It then calls the addUserXP function to add the specified amount of XP to the user in the guild. If the addition is successful, a success embed is sent as a reply, indicating the amount of XP added and the user's updated XP. If the addition fails, an error embed is sent, indicating that the user has no XP or an error occurred.